back - set specific excel column width
